:sparkles:feat: linkedContentId 타입 Integer->Long으로 변경

+++ src/main/java/com/opu/opube/feature/notification/command/domain/aggregate/Notification.java
@@ -32,7 +32,7 @@ public class Notification {
     private Boolean isRead = false;
 
     @Column(name = "linked_content_id")
-    private Integer linkedContentId;
+    private Long linkedContentId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id", nullable = false)
